requirements.txt
adal
astroid
azure-common
certifi
cffi
chardet
click
colorama
Flask
Flask-SQLAlchemy
idna
isodate
isort
itsdangerous
Jinja2
lazy-object-proxy
MarkupSafe
mccabe
msrest
msrestazure
oauthlib
psycopg2-binary
pycparser
PyJWT
pylint
python-dateutil
python-http-client
requests
requests-oauthlib
six
SQLAlchemy
toml
uamqp
urllib3
Werkzeug
wrapt
sendgrid
azure.servicebus
azure.functions
cryptography

